+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# NEXT
+# 3.8.4
 - `KeyboardConfig` improvements. Now `directionalKeys` expect list of `KeyboardDirectionalKeys`. Fix [#507](https://github.com/RafaelBarbosatec/bonfire/issues/507)
 - `PlatformEnemy` improvements.
 - Adds `flipAnimation` method in `ui.Image`.